+++ Assets/Animations/mario-beetle/beetleExplode.anim
@@ -22,13 +22,13 @@ AnimationClip:
     curve:
     - time: 0
       value: {fileID: -2234714183474105036, guid: a05bd78e6765dcf45af918ebaddb7ffc, type: 3}
-    - time: 0.016666668
+    - time: 0.16666667
       value: {fileID: 2860039434169806805, guid: a05bd78e6765dcf45af918ebaddb7ffc, type: 3}
-    - time: 0.033333335
+    - time: 0.33333334
       value: {fileID: -8137531386856967648, guid: a05bd78e6765dcf45af918ebaddb7ffc, type: 3}
-    - time: 0.05
+    - time: 0.5
       value: {fileID: 429100700143574792, guid: a05bd78e6765dcf45af918ebaddb7ffc, type: 3}
-    - time: 0.06666667
+    - time: 0.6666667
       value: {fileID: 1267864533, guid: a05bd78e6765dcf45af918ebaddb7ffc, type: 3}
     attribute: m_Sprite
     path: 
@@ -62,7 +62,7 @@ AnimationClip:
     m_AdditiveReferencePoseClip: {fileID: 0}
     m_AdditiveReferencePoseTime: 0
     m_StartTime: 0
-    m_StopTime: 0.083333336
+    m_StopTime: 0.68333334
     m_OrientationOffsetY: 0
     m_Level: 0
     m_CycleOffset: 0
