+++ Assets/KunaiProjectile.cs
@@ -23,6 +23,11 @@ public class KunaiProjectile : MonoBehaviour
             FlyMovement enemy = collision.GetComponent<FlyMovement>();
             enemy.OnHit(1);
         }
+        else if (collision.CompareTag("MarioBeetle"))
+        {
+            BeetleMovementController enemy = collision.GetComponent<BeetleMovementController>();
+            enemy.OnHit(1);
+        }
         Destroy(gameObject);
     }
 }
