+++ Assets/PhysicsMaterial/BeetleTop.physicsMaterial2D
@@ -1,0 +1,14 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!62 &6200000
+PhysicsMaterial2D:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_Name: BeetleTop
+  serializedVersion: 2
+  friction: 1
+  bounciness: 0
+  m_FrictionCombine: 1
+  m_BounceCombine: 4
