+++ Assets/Prefabs/BeetleEnemyWalkable.prefab
@@ -77,7 +77,7 @@ GameObject:
   - component: {fileID: -8229421698484948710}
   - component: {fileID: 9179391039753524795}
   m_Layer: 9
-  m_Name: BeetleEnemy
+  m_Name: BeetleEnemyWalkable
   m_TagString: MarioBeetle
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
