+++ Assets/Prefabs/BeetleEnemyWalkable.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 3a6ef9fc12863674293fd17b643178dd
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
