+++ Assets/Prefabs/BeetleGooExplosion.prefab
@@ -75,8 +75,8 @@ SpriteRenderer:
   m_AutoUVMaxDistance: 0.5
   m_AutoUVMaxAngle: 89
   m_LightmapParameters: {fileID: 0}
-  m_SortingLayerID: 0
-  m_SortingLayer: 0
+  m_SortingLayerID: -966137473
+  m_SortingLayer: 5
   m_SortingOrder: 0
   m_Sprite: {fileID: -2234714183474105036, guid: a05bd78e6765dcf45af918ebaddb7ffc, type: 3}
   m_Color: {r: 1, g: 1, b: 1, a: 1}
