+++ Assets/Scripts/WalkableBeetleScript.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 543cdec2ca7a72641a74869c08b9a0bf
